+++ .pax/pre-packaging.sh
@@ -14,6 +14,10 @@ set -x
 FUNC=[CreatePax][pre-packaging]
 PWD=$(pwd)
 
+cd ./content/server
+npm install --production
+cd ${PWD}
+
 # display extracted files
 echo "$FUNC content of $PWD...."
 find . -print