+++ tests/UnitTests/testResources/reducers/jobNodesResources.js
@@ -5,7 +5,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *
- * Copyright IBM Corporation 2018, 2019
+ * Copyright IBM Corporation 2018, 2020
  */
 
 import { Map, List } from 'immutable';
@@ -30,6 +30,7 @@ export const receivedJobsState = Map({
                 returnCode: 'CC 0000',
                 status: 'OUTPUT',
                 isToggled: false,
+                isSelected: false,
                 files: List([]),
             }),
             Map({
@@ -39,6 +40,7 @@ export const receivedJobsState = Map({
                 returnCode: 'ABEND S222',
                 status: 'OUTPUT',
                 isToggled: false,
+                isSelected: false,
                 files: List([]),
             }),
             Map({
@@ -48,6 +50,7 @@ export const receivedJobsState = Map({
                 returnCode: 'ABEND S222',
                 status: 'OUTPUT',
                 isToggled: false,
+                isSelected: false,
                 files: List([]),
             }),
         ],
@@ -65,6 +68,7 @@ export const toggledJobState = Map({
                 returnCode: 'CC 0000',
                 status: 'OUTPUT',
                 isToggled: true,
+                isSelected: false,
                 files: List([]),
             }),
             Map({
@@ -74,6 +78,7 @@ export const toggledJobState = Map({
                 returnCode: 'ABEND S222',
                 status: 'OUTPUT',
                 isToggled: false,
+                isSelected: false,
                 files: List([]),
             }),
             Map({
@@ -83,6 +88,7 @@ export const toggledJobState = Map({
                 returnCode: 'ABEND S222',
                 status: 'OUTPUT',
                 isToggled: false,
+                isSelected: false,
                 files: List([]),
             }),
         ],
@@ -100,6 +106,101 @@ export const receivedJobsStateSingleJobName = Map({
                 returnCode: 'CC 0000',
                 status: 'OUTPUT',
                 isToggled: false,
+                isSelected: false,
+                files: List([]),
+            }),
+        ],
+    ),
+    isFetching: false,
+});
+
+export const selectedSignleJobState = Map({
+    jobs: List(
+        [
+            Map({
+                jobName: 'TEST',
+                jobId: 'JOB1234',
+                label: 'TEST:JOB1234',
+                returnCode: 'CC 0000',
+                status: 'OUTPUT',
+                isToggled: false,
+                isSelected: true,
+                files: List([]),
+            }),
+        ],
+    ),
+    isFetching: false,
+});
+
+export const oneSelectedJobs = Map({
+    jobs: List(
+        [
+            Map({
+                jobName: 'TEST',
+                jobId: 'JOB1234',
+                label: 'TEST:JOB1234',
+                returnCode: 'CC 0000',
+                status: 'OUTPUT',
+                isToggled: false,
+                isSelected: true,
+                files: List([]),
+            }),
+            Map({
+                jobName: 'TEST1',
+                jobId: 'TSU18689',
+                label: 'TEST1:TSU18689',
+                returnCode: 'ABEND S222',
+                status: 'OUTPUT',
+                isToggled: false,
+                isSelected: false,
+                files: List([]),
+            }),
+            Map({
+                jobName: 'TEST1',
+                jobId: 'TSU18516',
+                label: 'TEST1:TSU18516',
+                returnCode: 'ABEND S222',
+                status: 'OUTPUT',
+                isToggled: false,
+                isSelected: false,
+                files: List([]),
+            }),
+        ],
+    ),
+    isFetching: false,
+});
+
+export const allSelectedJobs = Map({
+    jobs: List(
+        [
+            Map({
+                jobName: 'TEST',
+                jobId: 'JOB1234',
+                label: 'TEST:JOB1234',
+                returnCode: 'CC 0000',
+                status: 'OUTPUT',
+                isToggled: false,
+                isSelected: true,
+                files: List([]),
+            }),
+            Map({
+                jobName: 'TEST1',
+                jobId: 'TSU18689',
+                label: 'TEST1:TSU18689',
+                returnCode: 'ABEND S222',
+                status: 'OUTPUT',
+                isToggled: false,
+                isSelected: true,
+                files: List([]),
+            }),
+            Map({
+                jobName: 'TEST1',
+                jobId: 'TSU18516',
+                label: 'TEST1:TSU18516',
+                returnCode: 'ABEND S222',
+                status: 'OUTPUT',
+                isToggled: false,
+                isSelected: true,
                 files: List([]),
             }),
         ],
@@ -119,6 +220,7 @@ export const receivedJobFiles = Map({
                 returnCode: 'CC 0000',
                 status: 'OUTPUT',
                 isToggled: false,
+                isSelected: false,
                 files: [
                     {
                         label: 'JESMSGLG',